feat: add direct deposit migration base
