+++ src/applications/personalization/profile/components/Profile.jsx
@@ -164,7 +164,9 @@ class Profile extends Component {
     const toggles = this.props.profileToggles;
 
     const routes = getRoutes({
-      profileContactsPage: toggles.profileContacts,
+      profileContacts: toggles.profileContacts,
+      profileShowDirectDepositSingleForm:
+        toggles.profileShowDirectDepositSingleForm,
     });
 
     return (
