+++ src/applications/personalization/profile/constants.js
@@ -10,6 +10,9 @@ export const PROFILE_TOGGLES = {
   profileShowQuickSubmitNotificationSetting: false,
   profileShowEmailNotificationSettings: false,
   profileShowProofOfVeteranStatus: false,
+  profileShowDirectDepositSingleForm: false,
+  profileShowDirectDepositSingleFormAlert: false,
+  profileShowDirectDepositSingleFormEduDowntime: false,
 };
 
 // The values of these constants map to the possible values that come back from
