+++ src/applications/personalization/profile/tests/components/direct-deposit/PaymentHistory.unit.spec.jsx
@@ -3,7 +3,7 @@ import { expect } from 'chai';
 import sinon from 'sinon';
 import userEvent from '@testing-library/user-event';
 
-import PaymentHistory from '../../../components/direct-deposit/PaymentHistory';
+import PaymentHistory from '../../../components/direct-deposit/legacy/PaymentHistory';
 
 import { renderWithProfileReducers } from '../../unit-test-helpers';
