+++ src/applications/personalization/profile/tests/components/direct-deposit/alerts/DirectDepositBlocked.unit.spec.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { axeCheck } from 'platform/forms-system/test/config/helpers';
-import DirectDepositBlocked from '../../../../components/direct-deposit/alerts/DirectDepositBlocked';
+import DirectDepositBlocked from '../../../../components/direct-deposit/legacy/alerts/DirectDepositBlocked';
 
 describe('authenticated experience -- profile -- direct deposit', () => {
   describe('DirectDepositBlocked', () => {
